requirements.txt
discord.py
requests
sarge